requirements.txt
fastapi~=0.111.0
pydantic~=2.7.4
openai~=1.35.7
python-dotenv~=1.0.1
uvicorn~=0.30.1